Rename AV::Helpers::AssetPaths to AV::AssetPaths to solve autoload mess.
